+++ include/Config.hpp
@@ -41,6 +41,7 @@ struct Config
     float amplitude             = 1.0f;
     sonify::Direction direction = sonify::Direction::LEFT_TO_RIGHT;
     sf::ContextSettings window;
+    float image_rotation = 0.f;
     bool loop    = false;
     bool verbose = false;
 };